register with validation, check jwt to delete/update posts

+++ controllers/postController.js
@@ -3,7 +3,9 @@ const validaitonHandler = require("../validations/validationHandler");
 
 exports.index = async (req, res, next) => {
     try {
-        const posts = await Post.find().sort({ createdAt: -1 });
+        const posts = await Post.find()
+            .populate("user")
+            .sort({ createdAt: -1 });
         res.send(posts);
     } catch (err) {
         next(err);
@@ -14,7 +16,7 @@ exports.show = async (req, res, next) => {
     try {
         const post = await Post.findOne({
             _id: req.params.id,
-        });
+        }).populate("user");
         res.send(post);
     } catch (err) {
         next(err);
@@ -27,6 +29,7 @@ exports.store = async (req, res, next) => {
         let post = new Post();
         post.description = req.body.description;
         post.image = req.file.filename;
+        post.user = req.user;
         post = await post.save();
 
         res.send(post);
@@ -41,6 +44,16 @@ exports.update = async (req, res, next) => {
         let post = await Post.findById({
             _id: req.params.id,
         });
+        // Check if user created the specified post
+        if (!post || post.user != req.user.id) {
+            console.log(post.user);
+            console.log(req.user._id);
+            const error = new Error(
+                "You do not have permission to update this post."
+            );
+            error.statusCode = 400;
+            throw error;
+        }
         post.description = req.body.description;
         post = await post.save();
 
@@ -55,6 +68,16 @@ exports.delete = async (req, res, next) => {
         let post = await Post.findById({
             _id: req.params.id,
         });
+        // Check if user created the specified post
+        if (!post || post.user != req.user.id) {
+            console.log(post.user);
+            console.log(req.user._id);
+            const error = new Error(
+                "You do not have permission to delete this post."
+            );
+            error.statusCode = 400;
+            throw error;
+        }
         post = await post.delete();
 
         res.send({ message: `Post successfully deleted.` });
